fix: ajustes nas verificacoes do livro e cliente

+++ services/author.service.js
@@ -1,4 +1,5 @@
 import authorRepository from '../repositories/author.repository.js';
+import bookRepository from '../repositories/book.repository.js';
 
 async function createAuthor(author) {
     return await authorRepository.insertAuthor(author);
@@ -9,6 +10,12 @@ async function updateAuthor(author) {
 }
 
 async function deleteAuthor(author_id) {
+    const book = bookRepository.getBooksByAuthor(author_id);
+
+    if(book.length > 0) {
+        throw new Error('Existe livros para esse esse autor, não é possível excluir.');
+    }
+
     await authorRepository.deleteAuthor(author_id);
 }
